pyproject.toml
[project]
name = "orchestra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "chromadb>=1.1.0",
    "faiss-cpu>=1.12.0",
    "fastapi[standard]>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "lxml>=6.0.1",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sentence-transformers>=5.1.0",
    "supabase>=2.18.1",
    "tiktoken>=0.11.0",
    "transformers>=4.56.1",
]
